define limited 3 bond mapping (for no tp case)
define limited 3 bond mapping to rule out the measurements of unnecessary bonds for measuring Hall conductivity.
Notice that this is only for bps=2, which is the case for none-tp Hubbard model.

We can save the storage a little bit in this way. This is actually not so necessary but a minor improvement for saving the storage.