store the precauculation of limited 3 bond mapping
A comment for the laste comment:
Adding the limited 3 bond mapping does not save storage a lot.
Becasue in the hd5f files, the precaculation map_bbb_lim takes (num_b,num_b,num_b). which is the most storage taken.
But this big matrix is needed to help specify which measurements are necessary to speed up the measurement process.
So it is a necessary matrix to take this storage.