add the kernel for doing cubic spline intergral
Basically the idea is that no matter what you do, the end result is a linear function of the input. 
i.e.
G(iw_n) = sum_tau K(iw_n, tau) G(tau)

G(iw_n) and G(tau) are vectors. K(iw_n, tau) is a matrix.

Your goal is to determine the matrix, such that this matrix equation reproduces the exact same result, as if you did a spline fit, interpolated to a very fine tau grid, and then did a fourier transform the usual way. This is possible, because both the fourier transform and interpolation are linear operations.

The idea in that piece of code is to perform the entire procedure for G(tau) that look like
1,0,0,0,0,...
0,1,0,0,0,...
And then think of the actual G(tau) as a linear combination of those.

CubicSpline(taus, np.identity(L+1)) does the interpolation part for each of these in a single line

refer to appendix C of https://arxiv.org/pdf/1507.01012.pdf
(idea from the email communication with Edwin <edwinwhuang@gmail.com>)